% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_lastseg.R
\name{arima_lastseg_forecast}
\alias{arima_lastseg_forecast}
\title{ARIMA forecast on last segment}
\usage{
arima_lastseg_forecast(x, h, n_star = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{h}{horizon}
}
\description{
ARIMA forecast on last segment
}
