% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_models.R
\name{base_forecast_last_segment}
\alias{base_forecast_last_segment}
\title{Choose base model based on segment length}
\usage{
base_forecast_last_segment(x, h, n_star = 8)
}
\arguments{
\item{x}{numeric vector}

\item{h}{horizon}

\item{n_star}{threshold for GM(1,1) vs ARIMA}
}
\description{
Choose base model based on segment length
}
