% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_reg.R
\name{coherence_ok}
\alias{coherence_ok}
\title{Check coherence: top-level series equals sum of bottom-level series}
\usage{
coherence_ok(ytilde, S)
}
\arguments{
\item{ytilde}{reconciled forecasts (matrix)}

\item{S}{summing matrix}
}
\value{
TRUE if coherent, FALSE otherwise
}
\description{
Check coherence: top-level series equals sum of bottom-level series
}
