% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_bottomup.R
\name{detect_cp_bottomup}
\alias{detect_cp_bottomup}
\title{Bottom-up hierarchical changepoint detection}
\usage{
detect_cp_bottomup(
  Y,
  hierarchy,
  eta = 0,
  w = 20,
  soft_aggregate_window = 5,
  cost = "mean"
)
}
\arguments{
\item{Y}{T x p matrix}

\item{hierarchy}{hierarchy object}

\item{eta}{misalignment penalty (placeholder)}

\item{w}{window}

\item{soft_aggregate_window}{window around child cps to allow at parent}

\item{cost}{"mean" or "gaussian"}
}
\value{
list of cps per series
}
\description{
Bottom-up hierarchical changepoint detection
}
