% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_independent.R
\name{detect_cp_independent}
\alias{detect_cp_independent}
\title{Detect changepoints independently per series}
\usage{
detect_cp_independent(
  Y,
  levels,
  betas = NULL,
  w = 20,
  cost = c("mean", "gaussian")
)
}
\arguments{
\item{Y}{T x p matrix (columns = series)}

\item{levels}{integer vector of length p giving levels per series (0..L)}

\item{betas}{optional penalties per series}

\item{w}{window size}

\item{cost}{"mean" or "gaussian"}
}
\value{
list of integer vectors of changepoints per series
}
\description{
Detect changepoints independently per series
}
