% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_topdown.R
\name{detect_cp_topdown}
\alias{detect_cp_topdown}
\title{Top-down hierarchical changepoint detection}
\usage{
detect_cp_topdown(
  Y,
  hierarchy,
  eta = 0,
  w = 20,
  soft_inherit_window = 5,
  cost = "mean"
)
}
\arguments{
\item{Y}{T x p matrix}

\item{hierarchy}{hierarchy object}

\item{eta}{misalignment penalty weight (currently placeholder)}

\item{w}{window}

\item{soft_inherit_window}{integer; if >0, allow children to search near parent cps}

\item{cost}{"mean" or "gaussian"}
}
\value{
list of changepoints per series
}
\description{
Top-down hierarchical changepoint detection
}
