% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_partition.R
\name{detect_cp_window}
\alias{detect_cp_window}
\title{Windowed optimal partition (basic PELT-style) changepoint detection}
\usage{
detect_cp_window(
  x,
  w = 20,
  beta = NULL,
  cost_fun = segment_cost_mean,
  max_cp = Inf
)
}
\arguments{
\item{x}{numeric vector}

\item{w}{window size for candidate starts}

\item{beta}{penalty per changepoint (default log(n))}

\item{cost_fun}{function(x,a,b) returning segment cost}

\item{max_cp}{maximum number of changepoints to return}
}
\value{
integer vector of changepoint locations (end indices of segments)
}
\description{
Windowed optimal partition (basic PELT-style) changepoint detection
}
