% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_reg.R
\name{ensure_colnames}
\alias{ensure_colnames}
\title{Ensure column names of a matrix or data.frame}
\usage{
ensure_colnames(x, names)
}
\arguments{
\item{x}{matrix or data.frame}

\item{names}{vector of column names}
}
\description{
Ensure column names of a matrix or data.frame
}
