% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm11.R
\name{gm11_forecast}
\alias{gm11_forecast}
\title{Forecast GM(1,1)}
\usage{
gm11_forecast(fit, h, n_obs)
}
\arguments{
\item{fit}{model from gm11_fit}

\item{h}{horizon}

\item{n_obs}{number of observed points}
}
\description{
Forecast GM(1,1)
}
