% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{H_from_S}
\alias{H_from_S}
\title{Construct H (upper rows of S beyond bottom series)}
\usage{
H_from_S(S)
}
\arguments{
\item{S}{Summing matrix.}
}
\value{
Matrix H.
}
\description{
Construct H (upper rows of S beyond bottom series)
}
