% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{last_segment}
\alias{last_segment}
\title{Extract last segment given changepoints}
\usage{
last_segment(x, tau)
}
\arguments{
\item{x}{numeric vector}

\item{tau}{integer vector of cps}
}
\description{
Extract last segment given changepoints
}
