% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_series_cps}
\alias{plot_series_cps}
\title{Plot a series with detected changepoints}
\usage{
plot_series_cps(x, tau, title = "Series with changepoints")
}
\description{
Plot a series with detected changepoints
}
