% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_reg.R
\name{reconcile}
\alias{reconcile}
\title{Reconcile coherent forecasts from bottom forecasts}
\usage{
reconcile(yhatB, S)
}
\arguments{
\item{yhatB}{h x nb matrix of bottom forecasts}

\item{S}{summing matrix}
}
\description{
Reconcile coherent forecasts from bottom forecasts
}
