% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{segment_cost_gaussian}
\alias{segment_cost_gaussian}
\title{Segment cost: Gaussian log-variance}
\usage{
segment_cost_gaussian(x, a, b)
}
\description{
Segment cost: Gaussian log-variance
}
\keyword{internal}
