igi% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_hier_data}
\alias{sim_hier_data}
\title{Simulate a simple 2-level hierarchy with known cps}
\usage{
sim_hier_data(
  Tn = 120,
  nb = 3,
  cps_bottom = list(c(40, 90), c(60), c(70)),
  noise = 1
)
}
\arguments{
\item{Tn}{length of series}

\item{nb}{number of bottom series}
}
\description{
Simulate a simple 2-level hierarchy with known cps
}
